Renames packages and imports to new nomic.*
