+++ src/main/kotlin/nomic/api/TestingSpring.kt
@@ -1,4 +1,4 @@
-package nomic.game.manager.api
+package nomic.api
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
