+++ src/main/kotlin/nomic/domain/entities/User.kt
@@ -1,4 +1,4 @@
-package nomic.game.manager.domain.entities
+package nomic.domain.entities
 
 import com.fasterxml.jackson.annotation.JsonValue
